Leaderboard and History Screens

+++ app/src/main/java/com/example/quizapp/model/Result.kt
@@ -7,6 +7,6 @@ data class Result(
     val userId: String = "",
     val userEmail: String = "",
     val quizCategory: String = "",
-    val score: Int = 0,
+    val score: Long = 0,
     val totalQuestions: Int = 0,
 )
