+++ app/src/main/java/com/example/quizapp/model/SResult.kt
@@ -4,6 +4,6 @@ data class SResult(
     val userId: String? = "",
     val userEmail: String? = "",
     val quizCategory: String = "",
-    val score: Int = 0,
+    val score: Long = 0,
     val totalQuestions: Int = 0,
 )
